requirements.txt
flask==3.0.3
gunicorn==21.2.0
numpy==1.26.4
scipy==1.11.4
scikit-fuzzy==0.5.0
networkx==3.2.1
mysql-connector-python==8.3.0